fix sparse attn api
