+++ src/paddlefleet/tilelang_ops/__init__.py
@@ -44,7 +44,7 @@ def __getattr__(name):
         globals().update(exports)
         return exports[name]
     if name == "csa_sparse_attn":
-        from .compressed_sparse_attn import csa_sparse_attn
+        from paddlefleet.fusions.csa_sparse_attn import csa_sparse_attn
 
         globals()[name] = csa_sparse_attn
         return csa_sparse_attn
