+++ src/paddlefleet/transformer/csa_attention.py
@@ -518,77 +518,6 @@ def _apply_rope(
 # ---------------------------------------------------------------------------
 
 
-def unfused_compressed_sparse_attn(
-    query: Tensor,
-    kv_full: Tensor,
-    attn_sink: Tensor,
-    topk_indices: Tensor,
-    softmax_scale: float,
-) -> Tensor:
-    """Sparse attention with MQA and learnable attention sink.
-
-    Args:
-        query: [b, sq, np, hn] multi-head query
-        kv_full: [b, n_kv, hn] single-head KV (original + compressed concatenated)
-        attn_sink: [np] per-head learnable bias (attention sink)
-        topk_indices: [b, sq, topk] indices into kv_full dim=1 (-1 = invalid)
-        softmax_scale: attention scale factor
-
-    Returns:
-        output: [b, sq, np * hn]
-    """
-    b, sq, np_heads, hn = query.shape
-    topk = topk_indices.shape[-1]
-
-    # Clamp negative indices to 0 for gathering, mask them later
-    safe_indices = paddle.clip(topk_indices, min=0).cast(
-        paddle.int64
-    )  # [b, sq, topk]
-    safe_indices_exp = safe_indices.unsqueeze(-1).expand(
-        [-1, -1, -1, hn]
-    )  # [b, sq, topk, hn]
-
-    # Gather KV at selected positions: [b, n_kv, hn] -> [b, sq, topk, hn]
-    kv_gathered = paddle.gather(
-        kv_full.unsqueeze(1).expand([-1, sq, -1, -1]),
-        dim=2,
-        index=safe_indices_exp,
-    )
-    with paddle.amp.auto_cast(False):
-        # Compute attention scores: [b, np, sq, topk]
-        q = query.transpose([0, 2, 1, 3]).cast("float32")  # [b, np, sq, hn]
-        kv_g = kv_gathered.cast("float32")
-        scores = (
-            paddle.einsum("bnsh,bskh->bnsk", q, kv_g) * softmax_scale
-        )  # [b, np, sq, topk]
-        # Mask invalid positions (topk_indices < 0) with -inf
-        invalid_mask = (topk_indices < 0).unsqueeze(1)  # [b, 1, sq, topk]
-        scores = scores.masked_fill(invalid_mask, float("-inf"))
-
-        # Softmax with attention sink
-        # sink: [np] -> [1, np, 1, 1]
-        sink = attn_sink.reshape([1, np_heads, 1, 1])
-        # Compute stable softmax: max over scores and sink
-        scores_max = scores.max(axis=-1, keepdim=True)  # [b, np, sq, 1]
-        scores_max = paddle.maximum(scores_max, sink)
-
-        exp_scores = paddle.exp(scores - scores_max)  # [b, np, sq, topk]
-        exp_sink = paddle.exp(sink - scores_max)  # [b, np, sq, 1]
-
-        sum_exp = (
-            exp_scores.sum(axis=-1, keepdim=True) + exp_sink
-        )  # [b, np, sq, 1]
-        attn_weights = exp_scores / sum_exp  # [b, np, sq, topk]
-
-        # Weighted sum: [b, np, sq, topk] x [b, sq, topk, hn] -> [b, np, sq, hn]
-        output = paddle.einsum("bnsk,bskh->bnsh", attn_weights, kv_g)
-    output = output.cast(query.dtype)
-
-    # Reshape: [b, np, sq, hn] -> [b, sq, np * hn]
-    output = output.transpose([0, 2, 1, 3]).reshape([b, sq, np_heads * hn])
-    return output
-
-
 def _resolve_csa_indexer_loss_topk_effective(
     config, index_topk: int, n_compressed: int
 ) -> int:
@@ -2361,34 +2290,21 @@ class CompressedSparseAttention(FleetLayer):
         topk_idxs: Tensor,
         softmax_scale: float,
     ):
+        from paddlefleet.fusions.csa_sparse_attn import csa_sparse_attn
+
         attn_sink_fp32 = (
             attn_sink.cast("bfloat16").cast("float32")
             if _ACCURACY_COMPATIBLE_KERNEL
             else attn_sink.cast("float32")
         )
-        if _resolve_csa_tilelang_switch(
-            self.config,
-            "csa_tilelang_enable_sparse_attn",
-        ):
-            from paddlefleet.fusions.csa_sparse_attn import csa_sparse_attn
-
-            sparse_attn_backend = getattr(
-                self.config, "csa_sparse_attn_backend", "tilelang"
-            )
-            output = csa_sparse_attn(
-                query,
-                kv_full,
-                attn_sink_fp32,
-                topk_idxs,
-                softmax_scale,
-                backend=sparse_attn_backend,
-            )
-        else:
-            output = unfused_compressed_sparse_attn(
-                query,
-                kv_full,
-                attn_sink_fp32,
-                topk_idxs,
-                softmax_scale,
-            )
-        return output
+        sparse_attn_backend = getattr(
+            self.config, "csa_sparse_attn_backend", "cudnn"
+        )
+        return csa_sparse_attn(
+            query,
+            kv_full,
+            attn_sink_fp32,
+            topk_idxs,
+            softmax_scale,
+            backend=sparse_attn_backend,
+        )
