+++ tests/multi_card_tests/transformer/test_csa_attention_cp.py
@@ -166,7 +166,7 @@ def _build_csa_config(
         dsa_indexer_loss_coeff=dsa_indexer_loss_coeff,
         dsa_indexer_use_sparse_loss=False,
         csa_tilelang_enable_indexer=False,
-        csa_tilelang_enable_sparse_attn=False,
+        csa_sparse_attn_backend="unfused",
         init_method=None,
         init_method_std=0.02,
         layernorm_epsilon=1e-5,
@@ -482,7 +482,7 @@ class TestDSv4HybridAttentionCP(unittest.TestCase):
         config.dsa_indexer_loss_coeff = 0.0
         config.dsa_indexer_use_sparse_loss = False
         config.csa_tilelang_enable_indexer = False
-        config.csa_tilelang_enable_sparse_attn = False
+        config.csa_sparse_attn_backend = "unfused"
         config.init_method = None
         config.init_method_std = 0.02
         config.output_layer_init_method = None
@@ -956,7 +956,7 @@ class TestTileLangCSALayerCP(unittest.TestCase):
             dsa_indexer_loss_coeff=loss_coeff,
             dsa_indexer_use_sparse_loss=use_sparse_loss,
             csa_tilelang_enable_indexer=True,
-            csa_tilelang_enable_sparse_attn=False,
+            csa_sparse_attn_backend="unfused",
             csa_tilelang_backend="attention_paddle_compat",
             csa_indexer_backend="tilelang",
             init_method=None,
