+++ tests/single_card_tests/custom_ops/test_tilelang_csa_indexer.py
@@ -1835,7 +1835,7 @@ class TestCSAForwardTileLangFwdOnlyPath(unittest.TestCase):
             dsa_indexer_loss_coeff=0.0,
             dsa_indexer_use_sparse_loss=False,
             csa_tilelang_enable_indexer=True,
-            csa_tilelang_enable_sparse_attn=False,
+            csa_sparse_attn_backend="unfused",
             init_method=None,
             init_method_std=0.02,
             layernorm_epsilon=1e-5,
